Update: Split SwordState to SwordState and ShieldState.
