+++ Source/GameObject/Character/Player/State/PlayerCharacterSwordState.cpp
@@ -89,30 +89,6 @@ namespace PlayerCharacterState
 			if (!owner->IsPlayer()) return;
 		}
 
-		// 特殊攻撃ステート
-		void AttackSpecialState::Enter()
-		{
-			owner->SetAnimationSpeed(1.2f);
-			owner->SetAnimation(PlayerCharacter::Animation::ANIM_GUARD_SHIELD_START, false, 0.05f);
-			owner->SetAnimation(PlayerCharacter::Animation::ANIM_GUARD_SHIELD_CONTINUE, true);
-		}
-		void AttackSpecialState::Execute(float elapsedTime)
-		{
-			// 反重力
-			owner->StopFall();
-			owner->StopMove();
-
-			if (!owner->InputAttackSpecial())
-			{
-				owner->GetStateMachine()->ChangeState(static_cast<int>(PlayerCharacter::STATE::IDLE));
-			}
-		}
-		void AttackSpecialState::Exit()
-		{
-			owner->SetAnimationSpeed(1.0f);
-			owner->SetAnimation(PlayerCharacter::Animation::ANIM_GUARD_SHIELD_FINISH, false, 0.05f);
-		}
-
 		// スキル_1ステート
 		void Skill1State::Enter()
 		{
